Animation
git-svn-id: https://mediacenter-pc/svn/GPGPU/CUDA/RaytracingEngine/trunk@1177 c4de0f11-4f5c-764a-a4aa-d9e147179871

+++ OBJReader.cpp
@@ -35,7 +35,7 @@
 #include "Consts.h"
 #include "OBJReader.h"
 
-const int NB_MAX_FACES=NB_MAX_PRIMITIVES*0.9f; // Max number of faces
+const int NB_MAX_FACES=static_cast<int>(NB_MAX_PRIMITIVES*0.9f); // Max number of faces
 
 OBJReader::OBJReader(void)
 {
